[0.7.4-test] Abstracted colliders into a generic Component type.
This also involved rewriting the BoxCollider to work with this more abstract setup.
This new system should support any type of convex collider that someone (probably not me tbh) could bother to write.

I need to think if the complexity of supporting concave hitboxes is worthwhile or if I should add first-class support for multi-shape hitboxes using a collider that has a LinkedList<Collider> member.